Rename files to use plural naming as this matches the URLs we are using for the routes
